backend2.2/fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose==3.3.0
celery[redis]
requests
openai==0.28.0
datetime
httpx
